requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
supabase
python-dotenv
sendgrid
firebase-admin
python-jose
passlib[bcrypt]